Factor the subscriptions and the backoff tracking into their own files
