+++ cmd/runner/k8s.go
@@ -40,14 +40,22 @@ func initiateK8s(ctx context.Context, namespace string, cfgMap string, errorC ch
 	// The convention exists that the per machine configmap name is simply the hostname
 	podMap := os.Getenv("HOSTNAME")
 
+	// In the event that initializing the k8s listener fails we try once every 30 seconds to get it working
+	tick := time.NewTicker(30 * time.Second)
+	defer tick.Stop()
+
 	for {
-		// If k8s is specified we need to start a listener for lifecycle
-		// states being set in the k8s config map or within a config map
-		// that matches our pod/hostname
-		if err := runner.ListenK8s(ctx, *cfgNamespace, *cfgConfigMap, podMap, listeners.Master, errorC); err != nil {
-			logger.Warn("k8s monitoring offline", "error", err.Error())
+		select {
+		case <-tick.C:
+			// If k8s is specified we need to start a listener for lifecycle
+			// states being set in the k8s config map or within a config map
+			// that matches our pod/hostname
+			if err := runner.ListenK8s(ctx, *cfgNamespace, *cfgConfigMap, podMap, listeners.Master, errorC); err != nil {
+				logger.Warn("k8s monitoring offline", "error", err.Error())
+			}
+		case <-ctx.Done():
+			return
 		}
-		<-time.After(30 * time.Second)
 	}
 }
